feat(js): post conditions can refer to result with arrow notation
This enables postconditions in function definitions as well as spec() constructs to refer to the
result of the function by using a an arrow function instead of a boolean expression. This is
particulary useful for anonymous functions and imperative functions which can be confusing when used
in a postcondition.

When using the function name of the defined function in its post condition, it is now interpreted as
being called *after* the function returned. In order to refer to the result of the original call,
you should use an arrow function as your post condition.