perf(vcgen): translate simple propositions directly to SMT instead of boolean expression
JavaScript expressions used as pre-/post-condition, invariant or assertion should be translated
directly to logical propositions using native and/or/not connectives. Other expressions will still
be translated as terms returning boolean values. This makes generated verification conditions
simpler and therefore improves performance of the SMT solving.

Function specifications with spec() and purity annotations with pure() can only be used as part of a
simple proposition with connectives (and/or/not) and not as part of a complex boolean expression
such as a function call. This is necessary to ensure the correct quantifier polarity for quantifier
instantiation.