fix(js,smt): remove broken support for arrays and array expressions
There was a rudementary implementation for array expressions in the JavaScript input which were
represented as linked lists in the generated SMT. However, this was never really tested and is
likely broken, therefore any arrays should cause an explicit error. Arrays might get reintroduced
later with proper support. In the meantime, classes can be used to model lists.

array literals and array expressions not supported anymore