fix(vcgen): add limited support for calls and other heap references in class invariants
In order to make them easier to use, class invariants should hold independent of the current heap as
classes are themselves are immutable. However, this meant that function calls were not possible in a
class invariant. This patch makes it possible to use function calls and heap references in class
invariants by adding an explicit heap to both the access trigger as well as the
universally-quantified invariant pattern. However, the class invariant still has to hold independent
of the heap which is verified upon instantiation of the class. Currently, this limits the support
for verifying programs with instantiations but it generally allows the use of function calls and
heap references in class invariants.