test(classes): add more tests for class invariants with function calls
This adds new tests for verifying programs with classes and function calls in class invariants.
There are still limitations in terms on verifying the invariant for newly instantiated objects but
otherwise this allows use cases such as a map function for a List class that is parametrized by a
property. Additionally, this also adds a second test for verifying the length of a List that is
mapped over.