fix(assert): asserts should be assumed for proving postconditions
Assertions were previously treated as conditional exceptions. However, this meant that a failing
assertion would also affect the verification of other (potentially unrelated) post conditions. With
this fix, assertions within a function body are always assumed to hold. Of course, unverified
assertions will still be reported as usual, so the overall verification result does not change.