fix(smt): use set instead of sequence for obj fields to avoid z3 bugs
Avoid buggy sequences altogether and use set type for fields of objects. Unfortunately, this might cause the SMT solver to return objects with default field values (which corresponds to infinite fields). Theoretically, this could still be synthesized in JavaScript with Proxies but that would be somewhat ugly especially if counterexamples are exposed to users as part of the development environment. For now, it seems that the generated sets by z3 do not actually exploit the default field option. Therefore, converting these back to JavaScript does not result in infinite objects. As mentioned above, we might need to implement the Proxy idea if we do need such objects.