fix(vcgen): when violating func precondition, still assume postcondition
The right translation for spec is (R => pre(f, x) & post(f, x) => S) instead of (R => pre(f, x) & (R & post(f, x)) => S). This is was previously established in the Lean proof and now the implementation is changed accordingly. As a side effect, calling function without having a verified precondition still assumes the postcondition. I think this is intended as later parts of the code will throw fewer distracting errors which are originating from the call.