feat(smt): synthesize function counterexample based on SMT app function
Implemented function synthesis based on the interpretation of the appX function in the SMT output. In general, the synthesized function are a series of if-statments such that the function arguments are compared with simple JavaScript values and return (potentially complex) JavaScript expressions that correspond to values from the SMT output. Theoretically, this approach also works for variable-argument functions but this is not currently implemented.