feat(codegen): spec() dynamically enforeced in gen. code via wrapping
Simple function specifications are now enforced dynamically in generated code. Here, simple means that function specifications are either top-level or part of a conjunction. Specs that are part of a disjunction or in a negative position are difficult to dynamically enforce and are not supported yet. For each of these simple specs, an assignment statement is generated that wraps the target function values such that both the pre- and the post-condition is checked upon each call. The set of wrappers is stored alongside the function to prevent accumulating wrappers through repeated wrapping. It's important to note that the wrapping also needs to happen if the spec is part of an assertion that is usually skipped. For example, when verifying a function postcondition, the precondition is assumed and not checked. However, if the precondition contains a function spec, that spec still needs to result in wrapping to catch precondition violation of the function spec.