refactor(vcgen): generate test code during vcgen for code and assertions
Previously, the vcgen uses a dedicated visitor for translating assertion to logic while accumulating the test code. At the end, the test code is transformed by a separate visitor such that asserting a spec inserts a wrapper, etc. With this refactoring, the representation of JavaScript programs will have separate classes for logical terms, propositions, expressions and statements. Thereby, the vcgen can translate both actual code as well as assertions. Additionally, the vcgen will also synthesize and transform test code in the same pass. This makes it possible to add a call to the test code and insert fresh free variables for the call arguments during vcgen when encountering specs in a positive position. In particular, this applies to both the final part of a VC as well as any nested specs with the right polarity. In a sense, the translation of assertions, the vcgen and the generation of the test code based on the input program were previously done in three different passes (and in three different files). The refactoring merges these together such that one single pass generates both the VC as well as the test code. In the end, this will be a cleaner approach that allows correct wrapping and the insertion of synthetic calls. Also, I discovered a bug while refactoring: the non-recursive spec of the current function when verifying a function body or postcondition did not result in a wrapping. However, this might be necessary for the test body when synthesizing the counter-example for a unsatisfied precondition of a recursive call.