feat(vcgen): synthesize calls for asserted specs to provoke violation
The SMT solver output can generally be used for synthesizing function counterexamples. However, the counterexamples are only incorrect for certain inputs. These input values are also returned by the SMT solver but their names are automatically generated during QI whenever a universal quantifier in a positive position is lifted. Here, an ad-hoc solution for the call synthesis problem is implemented that works for dynamically enforceable (simple) specs. When generating testcode for a function spec assertion in a positive position, add a callback to the corresponding quantifier. When this quantifier gets lifted during QI, a function call is synthesized immediately after the wrapper is installed using free variables that will be replaced with values obtained from the model if a counterxample was returned by the SMT solver.