fix(codegen): gen. calls need to satisfy precondition to assume post
The arguments for synthesized calls are counter-examples according to the SMT solver. Interestingly, if a postcondition can be assumed without satisfying the precondition, this also means that the argument values returned by the SMT solver do not satisfy the precondition. This causes an assertion violation in the wrong code and may not be helpful for development. For now, the implemented solution is to ensure that a postcondition can only be assumed for input values that satisfy the precondition. Unfortunately, this means that code can cause two verification errors where the second one is just caused by the first one.