feat(vcgen): support for user-defined methods in class definitions
Added support for user-defined methods in class declarations. Following JavaScript specification, a call expression is considered a method invocation if the callee is a member expression such as “this.a” or “this[expr]”. For generated universal quantifiers, the “this” argument is treated similarly to other arguments. There is currently no support for super or static methods. While verifying method bodies, only non-recursive specs of other methods can be assumed. However, after method bodies are verified, the subsequent code can assume fully expanded/inlined method definitions. Also, in order to avoid scoping issues for nested spec expressions, unique names for each “this” argument of each spec is used.