feat(vcgen): preamble for built-in classes and funcs such as console.log
Adding a preamble to support built-in functions or classes such as “console.log(x)”. The preamble gets processed by a subclass of the VCGenerator such that no VCs are actually verified. Instead, functions and classes with methods are converted to an SMT proposition which can be assumed when verifying the user program.