fix(verifier): enforce z3 4.6 or higher to avoid string vcgen issues
Dropping support for z3 4.5. From now on, z3 4.6 or higher is required. This is mainly due to problems with string concatenation and string length but even z3 4.6 has string issues, so will probably update to 4.7 in the future.

dropping support for z3 4.5. in order to use esverify, at least z3 4.6 is required