feat(vcgen): support for string length and character access by index
Adding a preamble to support built-in functions or classes such as “console.log(x)”. The preamble gets processed by a subclass of the VCGenerator such that no VCs are actually verified. Instead, functions and classes with methods are converted to an SMT proposition which can be assumed when verifying the user program.