feat(js): array constructor creates new array (size arg not supported)
The array constructor can now be used in a similar way to the array literal notation to create new arrays, e.g. Array(1,2,3) or new Array(1,2,3) is equivalent to [1, 2, 3]. In JavaScript, a single integer argument actually creates an empty array of the given size. However, neither sparse arrays nor mutability are supported by esverify, so it will be treated as a single element array instead.