feat(vcgen): assume access trigger for member expressions in assertions
Changed the vcgen rules such that membership expressions also generate access triggers. In an assertion context, both function calls and field access are treated in an uninterpreted way. For function calls, this is primarily an advantage for higher-order functions as it’s not necessary to verify preconditions of function calls that are simply used as part of an equation in a pre- or post-condition. The equivalent for field access is the check for membership. However, this is not strictly necessary for invariants, so some programs would actually benefit from having an access trigger for each field access in an assertion context. Even if the trigger does not help, adding an additional trigger also does not cause any correctness problems. Again, this is not true for function calls as pre() and post() need to be available and those are only added for function calls in actual code.