fix(js): changed array invariant syntax to standard JavaScript a.every()
Previously, a decidated syntax every(arr, x => x > 0) was used to express array invariants in assertions. A more natural way to express these is with  arr.every(x => x > 0)  which is standard JavaScript code with the same semantics. As a downside, this now applies to all patterns  *.every(x => *) found in an assertion context. However, this restriction is very specific and potentially less limiting than the previous syntax.